+++ control.lua
@@ -21,7 +21,6 @@ require 'features.donator_messages'
 require 'features.train_saviour'
 require 'features.fish_market'
 require 'features.free_item_logging'
---require 'features.infinite_storage_chest'
 require 'features.nuke_control'
 require 'features.player_colors'
 require 'features.reactor_meltdown'
