+++ map_gen/misc/infinite_storage_chest.lua
